adapted the code to some changes in PYQT with python >= 3.10, where QSliders now only take integers in some methods not floats

+++ src/napari_storm/ChannelControls.py
@@ -77,13 +77,13 @@ class ChannelControls(QWidget):
         self.Breset.clicked.connect(self.reset)
 
         self.Slider_colormap_range = RangeSlider(parent=parent)
-        self.Slider_colormap_range.setRange(0.0, 100.0)
-        self.Slider_colormap_range.setValue((10.0, 90.0))
+        self.Slider_colormap_range.setRange(0, 100)
+        self.Slider_colormap_range.setValue((10, 90))
         self.Slider_colormap_range.valueChanged.connect(self.adjust_colormap_range)
 
         self.Slider_opacity = QSlider(Qt.Horizontal)
-        self.Slider_opacity.setRange(0.0, 100.0)
-        self.Slider_opacity.setValue(100.0)
+        self.Slider_opacity.setRange(0, 100)
+        self.Slider_opacity.setValue(100)
         self.Slider_opacity.hide()
         self.Slider_opacity.valueChanged.connect(self.adjust_z_color_encoding_opacity)
 
@@ -94,10 +94,10 @@ class ChannelControls(QWidget):
             items.append(cmap.name)
             pixmap = QPixmap(20, 20)
             color = QColor(
-                cmap.colors[1][0] * 255,
-                cmap.colors[1][1] * 255,
-                cmap.colors[1][2] * 255,
-                cmap.colors[1][3] * 255,
+                int(cmap.colors[1][0] * 255),
+                int(cmap.colors[1][1] * 255),
+                int(cmap.colors[1][2] * 255),
+                int(cmap.colors[1][3] * 255),
             )
             pixmap.fill(color)
             icons.append(QIcon(pixmap))
